+++ lib/ruuuby.rb
@@ -1,6 +1,8 @@
 require 'ruuuby/version'
 require 'ruuuby/class/nil'
 require 'ruuuby/class/bool'
+require 'ruuuby/class/ary'
+require 'ruuuby/class/int'
 require 'ruuuby/enum/emoji'
 
 # wip: *Ruuuby* is flavored modifications & extensions for increased quality of Ruby coding life