+++ spec/enum/emoji_spec.rb
@@ -4,10 +4,19 @@ RSpec.describe Ruuuby::Enum::Emoji do
     it 'Ruuuby::Enum responds to [Emoji]' do
       expect(Ruuuby::Enum.const_defined?('Emoji')).to eq(true)
     end
-    it 'with correct values' do
-      expect(subject::WARNING).to eq("⚠️")
-      expect(subject::POSITIVE).to eq("✅")
-      expect(subject::NEGATIVE).to eq("❌")
+    context 'with correct values' do
+      it 'emoji: warning' do
+        expect(subject::WARNING.to_s).to eq("⚠️")
+        expect(subject::WARNING.class).to eq(Symbol)
+      end
+      it 'emoji: positive' do
+        expect(subject::POSITIVE.to_s).to eq("✅")
+        expect(subject::POSITIVE.class).to eq(Symbol)
+      end
+      it 'emoji: negative' do
+        expect(subject::NEGATIVE.to_s).to eq("❌")
+        expect(subject::NEGATIVE.class).to eq(Symbol)
+      end
     end
   end
 end