+++ CHANGELOG.md
@@ -1,9 +1,31 @@
-#0.0.2
-- add func{bool?} to class{Object}
-#0.0.1
+# 0.0.3
+- modify {Ruuuby::Enum::Emoji} from strings to symbols
+- fix gemspec to reflect correct git URLs
+- add first update to README.md
+- add the following functions:
+
+| class  | func added    |
+| -----  | ------------- |
+| Object | int?          |
+| Object | ary?          |
+| Array  | remove_empty! |
+
+# 0.0.2
+- add the following functions:
+
+| class  | func added    |
+| -----  | ------------- |
+| Object | bool?         |
+
+# 0.0.1
 - add CHANGELOG.md
-- add func{empty?} to class{NilClass}
 - add emoji enum {Ruuuby::Enum::Emoji}
 - add bin/audit (for running rdoc, rspecs, and coverage reports)
-#0.0.0
+- add the following functions:
+
+| class    | func added    |
+| -----    | ------------- |
+| NilClass | empty?        |
+
+# 0.0.0
 - initial project creation